requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1
django-admin-inline-paginator==0.4.0
django-admin-interface==0.26.0
django-adminfilters==2.1.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.2.0
pip-autoremove==0.10.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.15
django-import-export==3.0.2